requirements.txt


Django==4.2.27
psycopg2-binary==2.9.9
python-decouple==3.8
dj-database-url==2.1.0
gunicorn==21.2.0
djangorestframework==3.14.0
djoser==2.1.0
Pillow==10.4.0
redis==5.0.1
django-cors-headers==4.3.0
python-telegram-bot==21.6
python-dotenv==1.0.1
httpx==0.27.2
beautifulsoup4==4.12.3
apscheduler==3.10.4
